requirements.txt
pyodbc
psycopg[binary]
azure-storage-blob